requirements.txt
langchain-core
langchain-openai
langchain>=0.3.0
openai>=1.50.0
python-dotenv
